[css-typed-om] Add support for grid properties.
This patch adds support for some grid properties.

We add a new <flex> CSS value object, which meant rebaselining all the tests.

We're failing some tests because the 'grid', 'grid-template' and
'grid-gap' shorthands don't serialize to anything [1], so Typed OM
can't create an unsupported style value from it. Not sure if this is
a bug...

[1] https://cs.chromium.org/chromium/src/third_party/WebKit/Source/core/css/StylePropertySerializer.cpp?type=cs&l=406

Bug: 820299
Change-Id: I8ff871fffed2977ca20a623bd283d7cbdc247870