Serialize font shorthand like any other shorthand
The font shorthand had a short-circuit in StylePropertySerializer, which
meant that it was never serialized as a shorthand for style.cssText,
citing reasons of web incompatibility. However, it appears that all
other major browsers serialize font as a shorthand instead of having
style.cssText expand to all properties. Remove this short-circuit to
match other browsers.

Bug: 827921
Change-Id: I4d522b41a6153808b6b8dab17831afdd8315fdc3